requirements.txt
discord.py
requests
bs4
pymongo
dnspython
pillow
numpy
matplotlib